api tests to api and festures to subdirs
