+++ pytest.ini
@@ -1,3 +1,2 @@
 [pytest]
-bdd_features_base_dir = features
 pythonpath = .
